refactor(gateway): isolate cloud reclaim lifecycle
